Started migrating to Flask - Reorganising file structure for Flask - Creating Flask app.py server - Started going through and changing links to url_for structure
