--- Underdog/assets/home.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Mini Racing Game</title>
-    <link rel="stylesheet" href="stylesheets/main.css">
-    <audio id="homeScreenAudio">
-        <source src="sounds/2021-08-16_-_8_Bit_Adventure_-_www.FesliyanStudios.com.mp3" type="audio/mpeg">
-    </audio>
-</head>
-
-<body>
-    <div style="float: left; z-index: 2;">
-        <h1 id="gameTitle">UNDERDOG</h1>
-        <!-- <h2>Game Modes:</h2> -->
-        <div>
-            <div>
-                <h3 class="gameModeNames">Classic</h3>
-                <p class="gameModeSubHeads">(Precision-based)</p>
-                <ul>
-                    <li class="gameDifficultySelections"><a href="classicEasy.html">Easy</a></li>
-                    <li class="gameDifficultySelections"><a href="classicMedium.html">Medium</a></li>
-                    <li class="gameDifficultySelections"><a href="classicHard.html">Hard</a></li>
-                </ul>
-            </div>
-
-            <div>
-                <h3 class="gameModeNames">Button Mash</h3>
-                <p class="gameModeSubHeads">(Speed-based)</p>
-                <ul>
-                    <li class="gameDifficultySelections"><a href="buttonMashEasy.html">Easy</a></li>
-                    <li class="gameDifficultySelections"><a href="buttonMashMedium.html">Medium</a></li>
-                    <li class="gameDifficultySelections"><a href="buttonMashHard.html">Hard</a></li>
-                </ul>
-            </div>
-        </div>
-    </div>
-    <div>
-        <div id="tagLine" style="float: right;">One afternoon at Royal Ascot, an unlikely new champion emerges...</div>
-    </div>
-    <img id="unlikelyChampion" src="images/dog.gif">
-    <img id="signPost" src="images/signPost.png">
-    <script src="script.js" onload="initialiseGame()"></script>
-</body>
-
-</html>
-
-</html>
