--- Underdog/assets/instructions.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <title>Instructions</title>
-        <link rel="stylesheet" href="stylesheets/main.css">
-    </head>
-    <body>
-        <h2>How To Play</h2>
-        <div>
-            <ol>
-                <li>A random button sequence will be generated, which you'll be able to see at the top of your screen.</li>
-                <li>Each time you complete the sequence your rider will move one place forward.</li>
-                <li>You have 5 lives. Every mistake costs you one life.</li>
-            </ol>
-        </div>
-        <a href="home.html">Back to home</a>
-        <script src="index.js"></script>
-    </body>
-    </html>
-</html>
