--- Underdog/assets/stylesheets/main.css
@@ -1,182 +1,0 @@
-.container {
-    width: 150px;
-    height: auto;
-    margin: auto;
-    position: fixed;
-    right: 30px;
-    bottom: 30px;
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    grid-template-rows: 1fr 1fr;
-    gap: 3% 2.5%;
-    grid-auto-flow: row;
-    grid-template-areas: ". upArrow ." "leftArrow downArrow rightArrow";
-}
-
-.container>div>img {
-    margin: auto;
-}
-
-.leftArrow {
-    grid-area: leftArrow;
-}
-
-.downArrow {
-    grid-area: downArrow;
-}
-
-.rightArrow {
-    grid-area: rightArrow;
-}
-
-.upArrow {
-    grid-area: upArrow;
-}
-
-#racerIcons>img {
-    width: auto;
-    display: block;
-}
-
-.botIcon {
-    height: 70px;
-}
-
-#playerIcon {
-    height: 40px;
-}
-
-.hearts {
-    height: 30px;
-    width: auto;
-}
-
-body {
-    font-family: 'Consolas', Courier, monospace;
-    padding: 30px;
-    font-size: 20px;
-    color: lightgreen;
-    font-weight: bold;
-    background-image: url('https://images-wixmp-ed30a86b8c4ca887773594c2.wixmp.com/f/01e12093-e706-40de-907c-2fd67ac52684/d6yyln0-f0eeaf16-1de3-4454-ab80-ecc1289516e4.png?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1cm46YXBwOjdlMGQxODg5ODIyNjQzNzNhNWYwZDQxNWVhMGQyNmUwIiwiaXNzIjoidXJuOmFwcDo3ZTBkMTg4OTgyMjY0MzczYTVmMGQ0MTVlYTBkMjZlMCIsIm9iaiI6W1t7InBhdGgiOiJcL2ZcLzAxZTEyMDkzLWU3MDYtNDBkZS05MDdjLTJmZDY3YWM1MjY4NFwvZDZ5eWxuMC1mMGVlYWYxNi0xZGUzLTQ0NTQtYWI4MC1lY2MxMjg5NTE2ZTQucG5nIn1dXSwiYXVkIjpbInVybjpzZXJ2aWNlOmZpbGUuZG93bmxvYWQiXX0.cQW-Rnb4FD4MsnP3Oo-TF4UFwIk2G0rYXH4XxRSDwwc');
-    overflow: hidden;
-    /* Hide scrollbars */
-}
-
-#outcome {
-    font-size: 3em;
-}
-
-#gameMode {
-    text-align: right;
-    margin-bottom: 0px;
-    margin-top: 0px;
-    font-size: 3em;
-}
-
-#difficulty {
-    text-align: right;
-    font-size: 2em;
-    color: white;
-}
-
-h1 {
-    font-size: 3em;
-    margin-top: 0px;
-}
-
-#raceTrack {
-    vertical-align: text-bottom;
-    margin: 30vh 0px;
-}
-
-#startButton {
-    position: fixed;
-    bottom: 30px;
-    left: 30px;
-}
-
-#startButton>a>img {
-    height: 50px;
-    width: auto;
-}
-
-#startButton>a>img:hover {
-    opacity: 70%;
-}
-
-#tagLine {
-    font-size: 2.5em;
-    text-align: right;
-    width: 700px;
-    color: lightgreen;
-}
-
-#gameTitle {
-    margin-top: 0px;
-    font-size: 4em;
-}
-
-#signPost {
-    height: 150px;
-    width: auto;
-    position: fixed;
-    bottom: 50vh;
-    right: 10vw;
-}
-
-@-webkit-keyframes move {
-    from {
-        left: -50%;
-    }
-    to {
-        left: 200%;
-    }
-}
-
-@keyframes move {
-    from {
-        left: -50%;
-    }
-    to {
-        left: 200%;
-    }
-}
-
-#unlikelyChampion {
-    bottom: 40vh;
-    height: 50px;
-    width: auto;
-    position: absolute;
-    -webkit-animation: move 20s;
-    animation: move 20s;
-    z-index: -1;
-    left: -100vh;
-    /* transition-duration: 1s; */
-}
-
-.gameModeNames {
-    color: white;
-    margin-bottom: 0px;
-}
-
-a:link {
-    color: lightgreen;
-  }
-  
-a:hover {
-    color: white;
-}
-
-a:visited {
-    color: lightgreen;
-}
-
-a:visited:hover {
-    color: white;
-}
-
-.gameModeSubHeads {
-    margin-top: 0px;
-    font-size: 1em;
-    font-weight: normal;
-}
