+++ templates/buttonMashMedium.html
@@ -1,0 +1,90 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Horse Racing Game</title>
+    <link rel="stylesheet" href="stylesheets/main.css">
+    <audio id="happyBark">
+        <source src="sounds/mixkit-happy-puppy-barks-741(3).mp3" type="audio/mpeg">
+    </audio>
+    <audio id="homeScreenAudio">
+        <source src="sounds/2021-08-16_-_8_Bit_Adventure_-_www.FesliyanStudios.com.mp3" type="audio/mpeg">
+    </audio>
+</head>
+
+<body>
+    <!--Places / Race Status-->
+
+    <div style="display: inline; float: left;">
+        <div>
+            <div id="outcome">Ready?</div>
+            <div>
+                <img class="hearts" id="life1" src="images/heart.png">
+                <img class="hearts" id="life2" src="images/heart.png">
+                <img class="hearts" id="life3" src="images/heart.png">
+                <img class="hearts" id="life4" src="images/heart.png">
+                <img class="hearts" id="life5" src="images/heart.png">
+            </div>
+            <div>
+                <div id="arrowSequence1" style="color: orange;">Error: Sequence Not Loaded</div>
+                <div id="arrowSequence2">Error: Sequence Not Loaded</div>
+                <div id="arrowSequence3">Error: Sequence Not Loaded</div>
+                <div id="arrowSequence4">Error: Sequence Not Loaded</div>
+            </div>
+        </div>
+        <!--Player-->
+        <div>
+            <div id="playerTime"></div>
+        </div>
+    </div>
+    <div style="display: inline; float: right;">
+        <h1 id="gameMode">Button Mash</h1>
+        <div id="difficulty">Medium</div>
+    </div>
+    <!--Racetrack-->
+    <div>
+        <div id="raceTrack" width="100%">
+            <div id="racerIcons">
+                <img id="playerIcon" src="images/dog.gif">
+                <img id="bot1Icon" class="botIcon" src="images/horse.gif">
+                <img id="bot2Icon" class="botIcon" src="images/horse.gif">
+                <img id="bot3Icon" class="botIcon" src="images/horse.gif">
+                <img id="bot4Icon" class="botIcon" src="images/horse.gif">
+            </div>
+        </div>
+    </div>
+    <!--Start-->
+    <div id="startButton">
+        <a href="home.html"><img src="images/start.jpg"></a>
+    </div>
+    <!--Arrows-->
+    <div id="arrowHUD">
+        <div class="container">
+            <div class="upArrow">
+                <img id="arrowUp" src="images/arrowUp.jpg" style="display: inherit;" height="100%" width="100%">
+                <img id="arrowUpPressed" src="images/arrowUpPressed.jpg" style="display: none;" height="100%" width="100%">
+            </div>
+            <div class="leftArrow">
+                <img id="arrowLeft" src="images/arrowLeft.jpg" style="display: inherit;" height="100%" width="100%">
+                <img id="arrowLeftPressed" src="images/arrowLeftPressed.jpg" style="display: none;" height="95%" width="95%">
+            </div>
+            <div class="downArrow">
+                <img id="arrowDown" src="images/arrowDown.jpg" style="display: inherit;" height="100%" width="100%">
+                <img id="arrowDownPressed" src="images/arrowDownPressed.jpg" style="display: none;" height="95%" width="95%">
+            </div>
+            <div class="rightArrow">
+                <img id="arrowRight" src="images/arrowRight.jpg" style="display: inherit;" height="100%" width="100%">
+                <img id="arrowRightPressed" src="images/arrowRightPressed.jpg" style="display: none;" height="95%" width="95%">
+            </div>
+        </div>
+    </div>
+
+    <script src="script.js " onload="initialiseGame() "></script>
+</body>
+
+</html>
+
+</html>
